Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    float n;
+    double n;
     printf("");
-    scanf("%f",n);
-    printf("You entered: %f",n);
+    scanf("%lf",n);
+    printf("You entered: %lf",n);
     return 0;
 }
